requirements.txt
aiofiles==0.7.0
anyio==3.3.3
asgiref==3.4.1
black==21.9b0
click==8.0.3
colorama==0.4.4
docxcompose==1.3.3
docxtpl==0.14.2
et-xmlfile==1.1.0
fastapi==0.70.0
h11==0.12.0
idna==3.3
Jinja2==3.0.2
lxml==4.8.0
MarkupSafe==2.0.1
modeltranslation==0.25
mypy-extensions==0.4.3
openpyxl==3.0.9
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.3
pydantic==1.8.2
python-docx==0.8.11
python-dotenv==0.20.0
python-multipart==0.0.5
regex==2021.10.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
thefuzz==0.19.0
tomli==1.2.1
tqdm==4.62.3
transliterate==1.10.2
typing-extensions==3.10.0.2
unoconv==0.9.0
uvicorn==0.15.0
xlsxtpl==0.2.1
